refactor(frequencySchemas): move frequency change schemas to a dedicated file for better organization
